[ao][fx] fixing rare bug with QuantizedGraphModule class
Pull Request resolved: https://github.com/pytorch/pytorch/pull/89893

when using PNP with functional linears (or other use cases with on the fly packing), the packedparams are created on the fly meaning the node.op is call_function rather than get_attr and node.target is a function rather than a str. This would cause the class to error unexpectedly since it expects a str.
ghstack-source-id: 174826244

Differential Revision: [D41578951](https://our.internmc.facebook.com/intern/diff/D41578951/)